+++ src/components/Modules/CardNumberInput/index.jsx
@@ -3,10 +3,7 @@ import styled from 'styled-components';
 import LabeledInput from '../../Atoms/LabeledInput';
 import InputWrapper from '../../Atoms/InputWrapper';
 import Input from '../../Atoms/Input';
-import validator from '../../../validation';
-import { numberRegex } from '../../../constant/regularExpression';
 import { CardNumberContext } from '../../../context/CardNumberContext';
-import useFocus from '../../../hooks/useFocus';
 
 const InputContainer = styled.div`
   display: flex;
@@ -17,43 +14,8 @@ const InputContainer = styled.div`
 `;
 
 function CardNumberInput() {
-  const {
-    orders,
-    numbers,
-    validations,
-    refs,
-    currentOrderRef,
-    setNumbers,
-    setValidations,
-  } = useContext(CardNumberContext);
-
-  const { focusPrevOrder } = useFocus({
-    validate: validator.validateCardNumber,
-    orders,
-    validations,
-    refs,
-    currentOrderRef,
-  });
-
-  const handleNumberChange = ({ target, nativeEvent: { data, inputType } }) => {
-    if ((numberRegex.test(data) || !data) && target.value.length <= 4) {
-      const order = target.name;
-      const newNumber = target.value;
-
-      updateNumbers(order, newNumber);
-      updateValidations(order, validator.validateCardNumber(newNumber));
-      focusPrevOrder(order, newNumber, inputType);
-      currentOrderRef.current = order;
-    }
-  };
-
-  const updateNumbers = (order, newNumber) => {
-    setNumbers(prevNumbers => ({ ...prevNumbers, [order]: newNumber }));
-  };
-
-  const updateValidations = (order, isValid) => {
-    setValidations(prevValidation => ({ ...prevValidation, [order]: isValid }));
-  };
+  const { numbers, validations, refs, handleNumberChange } =
+    useContext(CardNumberContext);
 
   return (
     <LabeledInput text="카드 번호">
