+++ src/components/Modules/SecurityNumberInput/index.jsx
@@ -2,9 +2,7 @@ import styled from 'styled-components';
 import { useContext } from 'react';
 import LabeledInput from '../../Atoms/LabeledInput';
 import Input from '../../Atoms/Input';
-import validator from '../../../validation';
 import InfoLabel from '../../Atoms/InfoLabel';
-import { numberRegex } from '../../../constant/regularExpression';
 import { SecurityNumberContext } from '../../../context/SecurityNumberContext';
 
 const Container = styled.div`
@@ -18,20 +16,9 @@ const InfoLabelContainer = styled.div`
 `;
 
 function SecurityNumberInput() {
-  const { number, validation, setNumber, setValidation } = useContext(
+  const { number, validation, onNumberChange } = useContext(
     SecurityNumberContext
   );
-
-  const onNumberChange = ({ target, nativeEvent: { data } }) => {
-    if (numberRegex.test(data) || !data) {
-      setNumber(target.value);
-      updateValidation(target.value);
-    }
-  };
-
-  const updateValidation = number => {
-    setValidation(validator.validateSecurityNumber(number));
-  };
 
   return (
     <Container>
