+++ src/context/CardOwnerContext.jsx
@@ -1,13 +1,27 @@
 import { createContext, useState } from 'react';
 import { PLACEHOLDER } from '../constant';
+import validator from '../../src/validation';
+import { englishRegex } from '../constant/regularExpression';
 
 const CardOwnerContext = createContext();
 
 function CardOwnerContextProvider({ children }) {
   const [name, setName] = useState('');
   const [validation, setValidation] = useState(false);
-
   const ownerNameString = name || PLACEHOLDER.NAME;
+
+  const onNameChange = ({ target, nativeEvent: { data } }) => {
+    if (englishRegex.test(data) || !data) {
+      const name = target.value.toUpperCase();
+
+      setName(name);
+      updateValidation(name);
+    }
+  };
+
+  const updateValidation = name => {
+    setValidation(validator.validateOwnerName(name));
+  };
 
   return (
     <CardOwnerContext.Provider
@@ -16,8 +30,7 @@ function CardOwnerContextProvider({ children }) {
         ownerNameString,
         validation,
         isValid: validation,
-        setName,
-        setValidation,
+        onNameChange,
       }}
     >
       {children}
