+++ src/context/PasswordContext.jsx
@@ -1,4 +1,7 @@
 import { createContext, useState, useRef } from 'react';
+import validator from '../../src/validation';
+import { numberRegex } from '../constant/regularExpression';
+import useFocus from '../hooks/useFocus';
 
 const PasswordContext = createContext();
 
@@ -15,6 +18,37 @@ function PasswordContextProvider({ children }) {
 
   const isValid = Object.values(validations).every(valid => valid);
 
+  const { focusPrevOrder } = useFocus({
+    validate: validator.validatePassword,
+    orders,
+    validations,
+    refs,
+    currentOrderRef,
+  });
+
+  const onPasswordChange = ({ target, nativeEvent: { data, inputType } }) => {
+    if (numberRegex.test(data) || !data) {
+      const order = target.name;
+      const newNumber = target.value;
+
+      updatePassword(order, newNumber);
+      updateValidations(order, newNumber);
+      focusPrevOrder(order, newNumber, inputType);
+      currentOrderRef.current = order;
+    }
+  };
+
+  const updatePassword = (order, number) => {
+    setPassword(prevPassword => ({ ...prevPassword, [order]: number }));
+  };
+
+  const updateValidations = (order, number) => {
+    setValidations(prevValidations => ({
+      ...prevValidations,
+      [order]: validator.validatePassword(number),
+    }));
+  };
+
   return (
     <PasswordContext.Provider
       value={{
@@ -24,8 +58,7 @@ function PasswordContextProvider({ children }) {
         isValid,
         refs,
         currentOrderRef,
-        setPassword,
-        setValidations,
+        onPasswordChange,
       }}
     >
       {children}
