+++ src/context/SecurityNumberContext.jsx
@@ -1,4 +1,6 @@
 import { createContext, useState } from 'react';
+import validator from '../../src/validation';
+import { numberRegex } from '../constant/regularExpression';
 
 const SecurityNumberContext = createContext();
 
@@ -6,14 +8,24 @@ function SecurityNumberContextProvider({ children }) {
   const [number, setNumber] = useState('');
   const [validation, setValidation] = useState(false);
 
+  const onNumberChange = ({ target, nativeEvent: { data } }) => {
+    if (numberRegex.test(data) || !data) {
+      setNumber(target.value);
+      updateValidation(target.value);
+    }
+  };
+
+  const updateValidation = number => {
+    setValidation(validator.validateSecurityNumber(number));
+  };
+
   return (
     <SecurityNumberContext.Provider
       value={{
         number,
         validation,
         isValid: validation,
-        setNumber,
-        setValidation,
+        onNumberChange,
       }}
     >
       {children}
